Simulate cages with the same starting conditions.

Simulate cages with the same starting conditions.

sim_cages(N_0, max_t, R, A, D_slope, D_inter, process_error, plant_mort_coefs,
  plant_death_age, plant_repl, n_reps, n_chains = 1, ...)

Arguments

N_0
Matrix of # aphids at time t=0. Should be matrix of size n_plants x n_lines.

max_t
Time steps per cage. Single integer.

R
Max growth rates per aphid line. Vector of length n_lines.

A
Density dependence per aphid line. Vector of length n_lines.

D_slope
Dispersal slope per aphid line. Vector of length n_lines.

D_inter
Dispersal intercept per aphid line. Vector of length n_lines.

process_error
SD of process error. Single double.

plant_mort_coefs
Coefficients (b0, then b1) for aphid mortality after plant starts dying.

plant_death_age
Number of days after which a plant starts dying.

plant_repl
Data frame containing two columns, the first indicating the date of replacement, the second indicating the plant that gets replaced on that date. There should be no repeat plants per date, but multiple plants can be replaced on the same date.

n_reps
Number of reps (i.e., cages) per chain.

n_chains
Number of chains to use. Note that values >1 will result in extra info being printed to the console. You can ignore this. Defaults to 1.

...
Other options passed to rstan::sampling().